requirements.txt
streamlit
pandas
numpy
plotly
alpha-vantage
requests
mysql-connector-python
bcrypt